requirements.txt
django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0